Добавила кнопки очистить и помощь.

+++ src/main/kotlin/com/github/kathybekh/doubleHashMap/view/DoubleHashingMapView.kt
@@ -2,12 +2,20 @@ package com.github.kathybekh.doubleHashMap.view
 
 import com.github.kathybekh.doubleHashMap.controller.DoubleHashingMapController
 import javafx.geometry.Insets
+import javafx.scene.control.Alert
+import javafx.scene.control.Alert.AlertType
+import javafx.scene.control.ButtonType
+import javafx.scene.control.Label
 import javafx.scene.control.TextField
-import javafx.scene.layout.*
+import javafx.scene.layout.Background
+import javafx.scene.layout.BackgroundFill
+import javafx.scene.layout.BorderPane
+import javafx.scene.layout.CornerRadii
 import javafx.scene.paint.Color
 import javafx.scene.text.Font
 import javafx.stage.FileChooser
 import tornadofx.*
+
 
 class DoubleHashingMapView : View() {
     private val controller: DoubleHashingMapController by inject()
@@ -15,8 +23,9 @@ class DoubleHashingMapView : View() {
     override val root = BorderPane()
     private var keyField: TextField by singleAssign()
     private var valueField: TextField by singleAssign()
+    private lateinit var statusLabel: Label
 
-    private var rows = controller.generateListOfRows().observable()
+    private var rows = controller.generateRows().observable()
 
     init {
         title = "Double Hashing Map"
@@ -25,6 +34,13 @@ class DoubleHashingMapView : View() {
             background = Background(
                 BackgroundFill(Color.WHITESMOKE, CornerRadii.EMPTY, Insets.EMPTY)
             )
+
+            top {
+                statusLabel = label("Enter key and value or press help.")
+                statusLabel.font = Font("Arial", 20.0)
+                statusLabel.textFill = Color.web("#1a237e")
+            }
+
             center {
                 stackpane {
                     tableview(rows) {
@@ -75,25 +91,45 @@ class DoubleHashingMapView : View() {
                             if (keyField.text != "" && valueField.text != "") {
                                 controller.add(keyField.text, valueField.text)
                                 controller.updateTable(rows)
+                                keyField.clear()
+                                valueField.clear()
+                                statusLabel.text = "Successful add."
                             } else {
-                                println("Not enough data! Enter key and value.")
+                                statusLabel.text = "Not enough data! Enter key and value."
                             }
                         }()
 
                         addButton("images/blue.png", "FIND")
                         {
                             val foundValue = controller.find(keyField.text)
-                            if (keyField.text != "" && foundValue != null) {
-                                println(foundValue)
+//                            INFORMATION
+                            val alert = Alert(AlertType.CONFIRMATION)
+                            alert.title = "Find Result"
+                            alert.headerText = if (keyField.text != "" && foundValue != null) {
+                                "$foundValue"
                             } else {
-                                println("The item you were looking for was not found!")
+                                "The item you were looking for was not found!"
                             }
+                            alert.showAndWait()
+                            keyField.clear()
+                            valueField.clear()
                         }()
 
                         addButton("images/orange.png", "DELETE") {
                             if (keyField.text != "") {
-                                controller.delete(keyField.text)
-                                controller.updateTable(rows)
+                                val alert = Alert(AlertType.CONFIRMATION)
+                                alert.title = "Delete Entry"
+                                alert.headerText = "Are you sure want to delete this entry?"
+                                val option = alert.showAndWait()
+                                if (option.get() == ButtonType.OK) {
+                                    controller.delete(keyField.text)
+                                    controller.updateTable(rows)
+                                    statusLabel.text = "Entry was delete."
+                                } else {
+                                    statusLabel.text = "Entry was not delete."
+                                }
+                                keyField.clear()
+                                valueField.clear()
                             }
                         }()
 
@@ -103,10 +139,30 @@ class DoubleHashingMapView : View() {
                                 arrayOf(FileChooser.ExtensionFilter("TEXT files (*.txt)", "*.txt"))
                             )
                             if (dir.isEmpty()) {
+                                statusLabel.text = "File was not selected."
                                 return@addButton
                             }
                             controller.readFromFile(dir.first())
                             controller.updateTable(rows)
+                            statusLabel.text = "File data was moved to table."
+                        }()
+
+                        addButton("images/red.png", "CLEAR") {
+                            val alert = Alert(AlertType.CONFIRMATION)
+                            alert.title = "Delete All Entries"
+                            alert.headerText = "Are you sure want to delete all entries?"
+                            val option = alert.showAndWait()
+                            if (option.get() == ButtonType.OK) {
+                                controller.clear()
+                                controller.updateTable(rows)
+                                statusLabel.text = "Table was clear."
+                            } else {
+                                statusLabel.text = "Table was not clear."
+                            }
+                        }()
+
+                        addButton("images/turquoise.png", "HELP") {
+                            HelpWindow().openWindow()
                         }()
                     }
